Reorganize demo code into dedicated demo/ directory
Implements feature-based organization by moving all demo-related code
into a single demo/ directory at the root level.

Changes:
- Created demo/ directory with clear ownership of demo feature
- Moved demo.ts → demo/index.ts
- Moved demo adapters to demo/adapters/:
  - auto-type-input.ts (automated typing for demo)
  - console-renderer.ts (console-based rendering)
  - noop-sound.ts (silent mode)
- Updated all imports to work from new location
- Added "demo" script to package.json

Benefits:
- Excellent cohesion: All demo code in one place
- Clear feature boundaries: Easy to see what comprises the demo
- Easier maintenance: Demo can be updated/extended as a unit
- Better discoverability: Demo feature visible at root level
- Potential for separate packaging if needed

The main game adapters remain in src/adapters/ as they are part of
the core application.
